✨ add `CurrencyCode` enums
There are a multitude of forex APIs in the wild
offering a wealth of forex information.
Tempting as it was to consume
one of these [APIs](https://openexchangerates.org/api/currencies.json),
I've chosen to use compile-time currency-enums.

Considerations:
- the forex-rates API deals with a small set of major currencies;
  representing this currency-set with enums seems a reasonable MVP,
  and greatly reduces the number of moving parts
- enums help sidestep currency-code validation issues
- access additional info via the associated `java.util.Currency`
- with careful enum-naming, Gson can deserialize the currency-codes
  in the `rates` element without having to use `@SerializedName`

In a typical iterative environment,
I'd expect new currencies to be introduced gradually.
The currency-enums could be grown
alongside the set of supported currencies.